Move Variables into its own namespace and add MutableVariables
